# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Ejemplo",
        "description": "Documentación de la API generada con Swagger para tu aplicación Laravel.",
        "contact": {
            "email": "soporte@tuapp.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/v1/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Inicio de sesión",
                "operationId": "18c25f0d2e88cb0b16bb401e49ca43e1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token de acceso generado"
                    },
                    "401": {
                        "description": "Credenciales inválidas"
                    }
                }
            }
        },
        "/v1/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registro de usuario",
                "operationId": "7c060e9f723a774c8da8d29ae02b01e2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "nickname": {
                                        "type": "string",
                                        "example": "user123"
                                    },
                                    "dni": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado con éxito"
                    },
                    "500": {
                        "description": "Error durante el registro"
                    }
                }
            }
        },
        "/v1/reports": {
            "post": {
                "tags": [
                    "Reports"
                ],
                "summary": "Crear un reporte",
                "operationId": "5d50213a8ede77869f7373a382d13c44",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "traffic_light_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Semáforo dañado"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "dañado"
                                    },
                                    "evidences": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "file_path": {
                                                    "type": "string",
                                                    "example": "/uploads/image.png"
                                                },
                                                "file_type": {
                                                    "type": "string",
                                                    "example": "image"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reporte creado con éxito"
                    },
                    "500": {
                        "description": "Error al crear el reporte"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/v1/reports/my": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Listar mis reportes",
                "operationId": "676b22b07f1df7dc2760ef82ab585a28",
                "responses": {
                    "200": {
                        "description": "Lista de reportes del usuario"
                    },
                    "401": {
                        "description": "Usuario no autenticado"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/v1/traffic-lights/nearby": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Semáforos cercanos",
                "operationId": "3e08a0213c6e6c64d12e3341325ac831",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "query",
                        "description": "Latitud actual del usuario",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "example": -12.0464
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "description": "Longitud actual del usuario",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "example": -77.0428
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "description": "Radio en kilómetros para buscar semáforos",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "example": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de semáforos cercanos"
                    },
                    "401": {
                        "description": "Usuario no autenticado"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Reports",
            "description": "Reports"
        }
    ]
}